backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
numpy
python-dotenv
pytest
httpx